Loan schedule

+++ database/migrations/2024_12_10_125905_create_payment_schedules_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('payment_schedules', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('loan_id'); // Foreign key referencing the loans table
+            $table->integer('loan_id'); // Foreign key referencing the loans table
             $table->integer('week'); // Week number
             $table->decimal('amount_due', 10, 2); // Amount due for the week (installment + interest)
             $table->date('due_date'); // Due date for this installment
